DEV-2420: Move models into sub-module.
